validation figure.
