+++ msresist/figures/figureM5.py
@@ -1,104 +1,58 @@
 """
-This creates Figure 5: Tumor infiltrating immune cells
+This creates Figure 4: STK11 analysis
 """
-
 import pickle
 import numpy as np
 import pandas as pd
 import seaborn as sns
 from sklearn.linear_model import LogisticRegressionCV
 from sklearn.preprocessing import StandardScaler
-from .common import subplotLabel, getSetup
-from .figureM3 import build_pval_matrix, calculate_mannW_pvals, plot_clusters_binaryfeatures
-from .figure2 import plotPCA, plotDistanceToUpstreamKinase
+from .figure2 import plotDistanceToUpstreamKinase
+from .figureM4 import plot_clusters_binaryfeatures, build_pval_matrix, calculate_mannW_pvals
 from ..logistic_regression import plotROC, plotClusterCoefficients
+from .common import subplotLabel, getSetup
 
 
 def makeFigure():
     """Get a list of the axis objects and create a figure"""
     # Get list of axis objects
-    ax, f = getSetup((14, 7), (2, 4), multz={0: 1})
-
-    # Set plotting format
-    sns.set(style="whitegrid", font_scale=1.2, color_codes=True, palette="colorblind", rc={"grid.linestyle": "dotted", "axes.linewidth": 0.6})
+    ax, f = getSetup((12, 7), (2, 3), multz={0: 1, 4:1})
 
     # Add subplot labels
     subplotLabel(ax)
 
-    # Import DDMC clusters
+    # Phosphoproteomic aberrations associated with molecular signatures
+    sns.set(style="whitegrid", font_scale=1.2, color_codes=True, palette="colorblind", rc={"grid.linestyle": "dotted", "axes.linewidth": 0.6})
+
+    # Load Clustering Model from Figure 2
     with open('msresist/data/pickled_models/binomial/CPTACmodel_BINOMIAL_CL24_W15_TMT2', 'rb') as p:
         model = pickle.load(p)[0]
 
+    # Import Genotype data
+    mutations = pd.read_csv("msresist/data/MS/CPTAC/Patient_Mutations.csv")
+    mOI = mutations[["Sample.ID"] + list(mutations.columns)[45:54] + list(mutations.columns)[61:64]]
+    y = mOI[~mOI["Sample.ID"].str.contains("IR")]
+
+    # Find centers
     X = pd.read_csv("msresist/data/MS/CPTAC/CPTAC-preprocessedMotfis.csv").iloc[:, 1:]
     centers = pd.DataFrame(model.transform())
     centers.columns = np.arange(model.ncl) + 1
     centers["Patient_ID"] = X.columns[4:]
-    centers = centers.loc[~centers["Patient_ID"].str.endswith(".N"), :].sort_values(by="Patient_ID").set_index("Patient_ID")
-
-    # Import Cold-Hot Tumor data
-    y = pd.read_csv("msresist/data/MS/CPTAC/Hot_Cold.csv").dropna(axis=1).sort_values(by="Sample ID")
-    y = y.loc[~y["Sample ID"].str.endswith(".N"), :].set_index("Sample ID")
-    l1 = list(centers.index)
-    l2 = list(y.index)
-    dif = [i for i in l1 + l2 if i not in l1 or i not in l2]
-    centers = centers.drop(dif)
-
-    # Transform to binary
-    y = y.replace("Cold-tumor enriched", 0)
-    y = y.replace("Hot-tumor enriched", 1)
-    y = np.squeeze(y)
-
-    # Remove NAT-enriched samples
-    centers = centers.drop(y[y == "NAT enriched"].index)
-    y = y.drop(y[y == "NAT enriched"].index).astype(int)
-    assert all(centers.index.values == y.index.values), "Samples don't match"
-
-    # Normalize
-    centers = centers.T
-    centers.iloc[:, :] = StandardScaler(with_std=False).fit_transform(centers.iloc[:, :])
-    centers = centers.T
+    centers = centers.set_index("Patient_ID")
 
     # Hypothesis Testing
-    centers["TIIC"] = y.values
-    pvals = calculate_mannW_pvals(centers, "TIIC", 1, 0)
+    assert np.all(y['Sample.ID'] == centers.index)
+    centers["STK11"] = y["STK11.mutation.status"].values
+    pvals = calculate_mannW_pvals(centers, "STK11", 1, 0)
     pvals = build_pval_matrix(model.ncl, pvals)
-    plot_clusters_binaryfeatures(centers, "TIIC", ["CTE", "HTE"], ax[0], pvals=pvals)
+    plot_clusters_binaryfeatures(centers, "STK11", ["WT", "Mutant"], ax[0], pvals=pvals)
+    ax[0].legend(loc='lower left')
 
-    # Logistic Regression
-    lr = LogisticRegressionCV(cv=7, solver="saga", max_iter=100000, tol=1e-4, n_jobs=-1, penalty="elasticnet", class_weight="balanced", l1_ratios=[0.4, 0.9])
-    plotROC(ax[1], lr, centers.iloc[:, :-1].values, y, cv_folds=4, title="ROC TIIC")
-    plotClusterCoefficients(ax[2], lr.fit(centers.iloc[:, :-1], y.values), title="TIIC")
-
-    # plot Upstream Kinases
-    plotDistanceToUpstreamKinase(model, [6, 17, 18, 20], ax[3], num_hits=3)
-
-    # Re-define centers
-    centers = pd.DataFrame(model.transform())
-    centers.columns = np.arange(model.ncl) + 1
-    centers["Patient_ID"] = X.columns[4:]
-    centers = centers.loc[~centers["Patient_ID"].str.endswith(".N"), :].sort_values(by="Patient_ID").set_index("Patient_ID")
-
-    # Import Cold-Hot Tumor data
-    y = pd.read_csv("msresist/data/MS/CPTAC/Hot_Cold.csv").dropna(axis=1).sort_values(by="Sample ID")
-    y = y.loc[~y["Sample ID"].str.endswith(".N"), :].set_index("Sample ID")
-    l1 = list(centers.index)
-    l2 = list(y.index)
-    dif = [i for i in l1 + l2 if i not in l1 or i not in l2]
-    centers = centers.drop(dif)
-
-    # Transform to binary
-    y = y.replace("Cold-tumor enriched", 0)
-    y = y.replace("Hot-tumor enriched", 1)
-    y = np.squeeze(y)
-
-    # Remove NAT-enriched samples
-    centers = centers.drop(y[y == "NAT enriched"].index)
-    y = y.drop(y[y == "NAT enriched"].index).astype(int)
+    # Reshape data (Patients vs NAT and tumor sample per cluster)
+    centers = centers.reset_index().set_index("STK11")
+    centers = find_patients_with_NATandTumor(centers, "Patient_ID", conc=True)
+    y = find_patients_with_NATandTumor(y.copy(), "Sample.ID", conc=False)
     assert all(centers.index.values == y.index.values), "Samples don't match"
-
-    # Select clusters changed by STK status (from figure M4)
-    coi = [1, 5, 7, 9, 11, 12, 15, 19, 21, 22, 24]
-    centers = centers.loc[:, coi]
 
     # Normalize
     centers = centers.T
@@ -106,11 +60,44 @@ def makeFigure():
     centers = centers.T
 
     # Logistic Regression
-    lr = LogisticRegressionCV(cv=5, solver="saga", max_iter=100000, tol=1e-4, n_jobs=-1, penalty="l1", class_weight="balanced")
-    plotROC(ax[4], lr, centers.values, y, cv_folds=4, title="ROC TIIC")
-    plotClusterCoefficients(ax[5], lr.fit(centers, y.values), xlabels=coi, title="TIIC")
+    centers["STK11"] = y["STK11.mutation.status"].values
+    lr = LogisticRegressionCV(Cs=10, cv=10, solver="saga", max_iter=100000, tol=1e-4, n_jobs=-1, penalty="l1", class_weight="balanced")
+    plotROC(ax[1], lr, centers.iloc[:, :-1].values, centers["STK11"], cv_folds=4, title="ROC STK11")
+    plotClusterCoefficients(ax[2], lr.fit(centers.iloc[:, :-1], centers["STK11"].values), list(centers.columns[:-1]), title="STK11")
+    ax[2].legend(loc='lower right', prop={'size':8})
 
     # plot Upstream Kinases
-    plotDistanceToUpstreamKinase(model, [11, 19, 21, 24], ax[6], num_hits=3)
+    plotDistanceToUpstreamKinase(model, [1, 8, 9, 14, 22], ax[3], num_hits=3)
 
     return f
+
+
+def merge_binary_vectors(y, mutant1, mutant2):
+    """Merge binary mutation status vectors to identify all patients having one of the two mutations"""
+    y1 = y[mutant1]
+    y2 = y[mutant2]
+    y_ = np.zeros(y.shape[0])
+    for binary in [y1, y2]:
+        indices = [i for i, x in enumerate(binary) if x == 1]
+        y_[indices] = 1
+    return pd.Series(y_)
+
+
+def find_patients_with_NATandTumor(X, label, conc=False):
+    """Reshape data to display patients as rows and samples (Tumor and NAT per cluster) as columns.
+    Note that to do so, samples that don't have their tumor/NAT counterpart are dropped."""
+    xT = X[~X[label].str.endswith(".N")].sort_values(by=label)
+    xN = X[X[label].str.endswith(".N")].sort_values(by=label)
+    l1 = list(xT[label])
+    l2 = [s.split(".N")[0] for s in xN[label]]
+    dif = [i for i in l1 + l2 if i not in l1 or i not in l2]
+    X = xT.set_index(label).drop(dif)
+    assert all(X.index.values == np.array(l2)), "Samples don't match"
+
+    if conc:
+        xN = xN.set_index(label)
+        xN.index = l2
+        xN.columns = [str(i) + "_N" for i in xN.columns]
+        X.columns = [str(i) + "_T" for i in X.columns]
+        X = pd.concat([X, xN], axis=1)
+    return X
